remove cast, there is no need for it

+++ logrotate/src/LogRotateHelper.ipp
@@ -453,7 +453,7 @@ bool LogRotateHelper::createCompressedFile(std::string file_name, std::string gz
 
     size_t N;
     while ((N = fread(buffer, 1, buffer_size, input)) > 0) {
-        gzwrite(output, buffer, static_cast<size_t>(N));
+        gzwrite(output, buffer, N);
     }
     if (gzclose(output) != Z_OK);
     if (fclose(input) != 0) {
